adding existing data tests, renaming RollupBlock to TransitionBatch, pulling logic out of TransitionBatchBuilder
