--- rollup/rollup_block_submitter.go
@@ -1,13 +1,0 @@
-package rollup
-
-type RollupBlockSubmitter interface {
-  submit(block *RollupBlock) error
-}
-
-type BlockSubmitter struct {}
-func NewBlockSubmitter() *BlockSubmitter {
-  return &BlockSubmitter{}
-}
-func (d *BlockSubmitter) submit(block *RollupBlock) error {
-  return nil
-}
